+++ go/bridge/rename_noreplace_apple.go
@@ -1,0 +1,17 @@
+//go:build darwin || ios
+
+package bridge
+
+import (
+	"errors"
+
+	"golang.org/x/sys/unix"
+)
+
+func renameNoReplace(oldPath, newPath string) error {
+	err := unix.RenamexNp(oldPath, newPath, unix.RENAME_EXCL)
+	if errors.Is(err, unix.ENOTSUP) {
+		return errNoReplaceUnsupported
+	}
+	return err
+}